requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
streamlit==1.44.1
python-dotenv==1.1.0
requests==2.32.3
PyPDF2==3.0.1
pydantic==2.11.3
google-generativeai==0.8.5
python-multipart==0.0.20
